+++ src/backend/draw.rs
@@ -18,7 +18,8 @@ impl Backend {
         draw_config: &DrawConfig,
     ) -> Result<(), ErrDontCare> {
         let s = &mut self.state;
-        s.update_program(self.program.id, self.program.vao);
+        s.update_program(self.program.id);
+        s.update_vao(self.program.vao);
         s.update_blend_mode(draw_config.blend_mode);
         s.update_framebuffer(target_framebuffer);
         s.update_texture(object_texture.id);
@@ -42,8 +43,9 @@ impl Backend {
         Ok(())
     }
 
-    pub fn draw_line(
+    pub fn debug_draw(
         &mut self,
+        rectangle: bool,
         target_framebuffer: GLuint,
         target_dimensions: (u32, u32),
         from: (i32, i32),
@@ -51,20 +53,21 @@ impl Backend {
         color: (f32, f32, f32, f32),
     ) -> Result<(), ErrDontCare> {
         let s = &mut self.state;
-        s.update_program(self.lines_program.id, self.lines_program.vao);
+        s.update_program(self.debug_program.id);
+        s.update_vao(self.debug_program.vao[rectangle as usize]);
         s.update_framebuffer(target_framebuffer);
         s.update_viewport_dimensions(target_dimensions);
         s.disable_depth();
-        s.update_debug_line_color(color);
+        s.update_debug_color(color);
         let data = (
             (from.0 as f32 + 0.5) / target_dimensions.0 as f32 * 2.0 - 1.0,
             (from.1 as f32 + 0.5) / target_dimensions.1 as f32 * 2.0 - 1.0,
-            (to.0 as f32 + 0.5) / target_dimensions.0 as f32 * 2.0 - 1.0,
-            (to.1 as f32 + 0.5) / target_dimensions.1 as f32 * 2.0 - 1.0,
+            (to.0 as f32 + 0.75) / target_dimensions.0 as f32 * 2.0 - 1.0,
+            (to.1 as f32 + 0.75) / target_dimensions.1 as f32 * 2.0 - 1.0,
         );
-        s.update_debug_line_start_end(data);
+        s.update_debug_start_end(data);
         unsafe {
-            gl::DrawArrays(gl::LINES, 0, 2);
+            gl::DrawArrays(gl::LINE_STRIP, 0, if rectangle { 5 } else { 2 });
         }
 
         Ok(())