+++ src/lib.rs
@@ -53,13 +53,23 @@ pub trait DrawTarget {
     ) -> Result<(), ErrDontCare>;
 
     /// Draws a line from `from` to `to`.
-    ///
-    /// The pixel at `to` is not part of the line.
     fn receive_line(
         &mut self,
         ctx: &mut Context,
         from: (i32, i32),
         to: (i32, i32),
+        color: (f32, f32, f32, f32),
+    ) -> Result<(), ErrDontCare>;
+
+    /// Draws the bounding box of an axis-aligned rectangle specified by
+    /// its `lower_left` and `upper_right` corner.
+    ///
+    /// In case `lower_left` is to the right or above `upper_right`, the two points will be flipped.
+    fn receive_rectangle(
+        &mut self,
+        ctx: &mut Context,
+        lower_left: (i32, i32),
+        upper_right: (i32, i32),
         color: (f32, f32, f32, f32),
     ) -> Result<(), ErrDontCare>;
 }
@@ -91,6 +101,16 @@ impl<T: DrawTarget> DrawTarget for &mut T {
         color: (f32, f32, f32, f32),
     ) -> Result<(), ErrDontCare> {
         <T>::receive_line(self, ctx, from, to, color)
+    }
+
+    fn receive_rectangle(
+        &mut self,
+        ctx: &mut Context,
+        lower_left: (i32, i32),
+        upper_right: (i32, i32),
+        color: (f32, f32, f32, f32),
+    ) -> Result<(), ErrDontCare> {
+        <T>::receive_rectangle(self, ctx, lower_left, upper_right, color)
     }
 }
 
@@ -206,11 +226,10 @@ impl Context {
 
     /// Draws the a line going from `from` to `to` onto `target` with the given `color`.
     ///
-    /// The pixel at `to` is not part of the line.
     /// To draw this line to the window, use [`Context::window_surface`] as a target.
     ///
     /// [`Context::window_surface`]: struct.Context.html#method.window_surface
-    pub fn draw_line<T>(
+    pub fn debug_line<T>(
         &mut self,
         target: &mut T,
         from: (i32, i32),
@@ -221,6 +240,27 @@ impl Context {
         T: DrawTarget,
     {
         target.receive_line(self, from, to, color)
+    }
+
+    /// Draws the bounding box of an axis-aligned rectangle specified by
+    /// its `lower_left` and `upper_right` corner.
+    ///
+    /// In case `lower_left` is to the right or above `upper_right`, the two points will be flipped.
+    ///
+    /// To draw this rectangle to the window, use [`Context::window_surface`] as a target.
+    ///
+    /// [`Context::window_surface`]: struct.Context.html#method.window_surface
+    pub fn debug_rectangle<T>(
+        &mut self,
+        target: &mut T,
+        lower_left: (i32, i32),
+        upper_right: (i32, i32),
+        color: (f32, f32, f32, f32),
+    ) -> Result<(), ErrDontCare>
+    where
+        T: DrawTarget,
+    {
+        target.receive_rectangle(self, lower_left, upper_right, color)
     }
 
     /// Clears the given [`DrawTarget`], setting each pixel to `color`
@@ -340,7 +380,19 @@ impl DrawTarget for WindowSurface {
         color: (f32, f32, f32, f32),
     ) -> Result<(), ErrDontCare> {
         let dim = ctx.backend.window_dimensions();
-        ctx.backend.draw_line(0, dim, from, to, color)
+        ctx.backend.debug_draw(false, 0, dim, from, to, color)
+    }
+
+    fn receive_rectangle(
+        &mut self,
+        ctx: &mut Context,
+        lower_left: (i32, i32),
+        upper_right: (i32, i32),
+        color: (f32, f32, f32, f32),
+    ) -> Result<(), ErrDontCare> {
+        let dim = ctx.backend.window_dimensions();
+        ctx.backend
+            .debug_draw(true, 0, dim, lower_left, upper_right, color)
     }
 }
 
@@ -544,8 +596,33 @@ impl DrawTarget for Texture {
     ) -> Result<(), ErrDontCare> {
         let target = self.prepare_as_draw_target(ctx)?;
 
-        ctx.backend
-            .draw_line(target.frame_buffer_id, target.dimensions, from, to, color)
+        ctx.backend.debug_draw(
+            false,
+            target.frame_buffer_id,
+            target.dimensions,
+            from,
+            to,
+            color,
+        )
+    }
+
+    fn receive_rectangle(
+        &mut self,
+        ctx: &mut Context,
+        lower_left: (i32, i32),
+        upper_right: (i32, i32),
+        color: (f32, f32, f32, f32),
+    ) -> Result<(), ErrDontCare> {
+        let target = self.prepare_as_draw_target(ctx)?;
+
+        ctx.backend.debug_draw(
+            true,
+            target.frame_buffer_id,
+            target.dimensions,
+            lower_left,
+            upper_right,
+            color,
+        )
     }
 }